split+moved charmap function to own file
